# GC improvements 5: `Store{Diff,Event}` optimizations